Format documentation site to reflect folder/file.jl structure, update various docstrings.
Although this is probably not desired for the end user, for the sake of both ease of
implementation and better understanding the way the code interconnects, the code has
been partially documented by folder/file.jl structure. This is subject to change,
and ultimately it may not be necessary for many of the lower level methods to be
shown to the user. More important/next focus will be to implement something
resembling a guide which includes examples of what parameters/method calls
can/should be.