Add user guides for before/after preliminary UI changes. Include all src files in documentation.
Crude guide implemented with @example blocks, need to find a better way to display code that doesn't show output or throw errors. Need to include color theme.
All src files included in a hierarchy mirroring folder structure that already exists (this creates a few folders with only one file and one include.jl, which is for now ignored for documentation).