Change freestream to the form `t -> (Ux(t), Uy(t))`
Previously, the freestream was represented as a struct with fields `Ux`,
`Uy`, and `inclination` (each functions of time). In general, only the
x and y components of velocity need to be stored. Inclination can be
accounted for before-hand. Instead of storing two functions of time, it
now stores a single function that returns a tuple of floats.

Note: none of the examples have been updated to reflect this yet.