Give quantity functions more options
Quantity functions now take a `CopyPreference` to specify whether a copy
should be returned (`AlwaysCopy`) or a view to the state is preferred
(`PreferView`).

Some quantity functions that apply over subdomains now take a `subgrids`
keyword argument to specify which subdomains to return the quantity
over. A `gridranges` function now returns the (x, y) grid ranges for
each subdomain returned by a quantity function.

In order to reshape the returned quantities to match the grid, it was
necessary to add a `Grid` reference to the `IBState` struct.

Also, quantities for `x_velocity` and `y_velocity` were added.